requirements.txt
python-box
pyyaml
coloredlogs
numpy
matplotlib
pandas
seaborn
jupyter
coverage
json2html
mkdocs-with-pdf
mkdocs-material
mkdocs
mkdocstrings
plotly
xlrd==1.2.0
openpyxl
pandas
python-dotenv
requests